Add support for SET ACCESS METHOD in ALTER TABLE
This is a backport of PG15 commit: b0483263dda, motivating the feature
to swap in between heap, AO and AOCS tables.

Conflicts:
* make_new_heap() has extra params in GPDB
* The new description added to alter_table_type_to_string is not
backported as we have to backport 2ed532ee8c4, which will be quite
tedious.
* We maintain AlteredTableInfo in altertablenodes.h.
* ATSimplePermissions() has one less param in GPDB as we haven't yet
absorbed 2ed532ee8c4.
* AT_REWRITE_ACCESS_METHOD's value clashed w/ GPDB specific
AT_REWRITE_NEW_COLUMNS_ONLY_AOCS. Changed the mask for the latter.
* create_am_optimizer.out needed to be updated with the new cases.
* The newAccessMethod in AlteredTableInfo needed a read/out func
change. Its imperative that we serialize/deserialize it, otherwise it
doesn't get passed to the QE backends.

Original commit message follows:

The logic used to support a change of access method for a table is
similar to changes for tablespace or relation persistence, requiring a
table rewrite with an exclusive lock of the relation changed.  Table
rewrites done in ALTER TABLE already go through the table AM layer when
scanning tuples from the old relation and inserting them into the new
one, making this implementation straight-forward.

Note that partitioned tables are not supported as these have no access
methods defined.

Author: Justin Pryzby, Jeff Davis
Reviewed-by: Michael Paquier, Vignesh C
Discussion: https://postgr.es/m/20210228222530.GD20769@telsasoft.com