Fix a flaky test idle_in_transaction_session_timeout
The test gets stuck on ic-tcp because when session 2 does DECLARE
CURSOR the QD needs to set up interconnect with all QEs but one of
the QE is blocked on a 'suspend' injected fault and couldn't ack.
Note that this isn't an issue for ic-proxy even they share the same
code path (SetupTCPInterconnect()) because the acks will be handled
by the proxy processes instead of the backends.

Fix this by using 'sleep' fault injection instead. That also makes
it unnecessary to use multiple sessions. Move the test to regress
then.