Fix an issue where the pg_appendonly entry is not removed during AO->heap
During changing access method from AO to heap, we need to transfer the
pg_appendonly entry of the AO table to be associated with the newly
created temp heap table (TransferAppendonlyEntries). But we mistakenly
insert a new tuple instead of updating the existing tuple. So even the
one associated with the temp table is removed, we still end up having
the old pg_appendonly entry, which has the same oid as the resulting
heap table.

Fixing that. And add test cases for it too.