Tests of VACUUM (SKIP_LOCKED) (#14265)
On PG12 an option SKIP_LOCKED was added to VACUUM and ANALYZE to indicate
whether we need to skip tables when a lock cannot be acquired at the time.
However, relevant tests don't cover all the scenarios on GPDB. This PR is
to add more tests to cover the scenarios.