Bring back cdbsubselect_drop_distinct from 5X branch. (#14258)
In GP's convert_ANY_sublink_to_join() now, there is no simplification of
origin subquery, as a result sometimes the sublink cannot be pulled up,
which produced a bad query plan. In GP5, there is a subquery simplify
step which helped sublink pull-up, but we lose this simplification since
version 6X until master, which caused the sublink pull-up doesn't work as
only simple subquery is allowed to be pulled up. By this commit we brought
back cdbsubselect_drop_distinct and cdbsubselect_drop_orderby, then we can
got simplified subquery again and do the sublink pull-up optimization on it
which produced a better plan.