fix gp_gettmid to return correct startup timestamp. (#14204)
Previously, gp_gettmid() returns PgStartTime as the start time of gpdb  ("tmid") and saves it into instrument shared memory, but the value is wrong after converting to `TimestampTz` (which is int64). Since the value is saved in the shared memory, plus GPCC has been using 32-bit integer for tmid, adapting to int64 will require a lot of change in both GPDB's and GPCC's coed. Therefore, in this commit, gp_gettmid() converts PgStartTime to a correct int32 value instead.

Co-authored-by: Violet Cheng <vcheng@vmware.com>