[ORCA] Enable multiple grouping set specs (#14366)
After this commit, following produces valid plan in ORCA:
  ```
  CREATE TABLE t AS SELECT i as a, i as b FROM generate_series(1,3)i;
  EXPLAIN SELECT a, b FROM t GROUP BY a, ROLLUP(b);
  ```

In ORCA, multiple grouping set specs is implemented as the pairwise
concatenation of the individual elements from the different grouping
sets. In the above example, there are 2 sets:
  ```
  "a" that produces (1), (2), (3)
  "ROLLUP(b)" that produces (), (1), (2), (3).
  ```

Pairwise concatenation of the two sets produces groups:

  (1,),(2,),(3,),(1,1),(2,1),(3,1),(1,2),(2,2),(3,2),(1,3),(2,3),(3,3)

Based on the table contents and the groups, the query result produces:

  (1,),(2,),(3,),(1,1),(2,2),(3,3)

NB: Fallback if multiple grouping set specs reference duplicate alias

There is a known issue in upstream Postgres right now where planner
fails to distinguish between two duplicate alias columns where one
column is possibly nulled by the action of a grouping node. For example:
  ```
  SELECT i AS ai1, i AS ai2 FROM generate_series(1,3)i GROUP BY ai2, ROLLUP(ai1) ORDER BY ai1, ai2;
  ```

Above query produces a range table entry (RTE) with a single target list
entry. In other words it fails to distinguish between ai1 and ai2. This
blunder can lead to wrong results.

So, you may ask, why should we fallback to PLANNER that can lead to
wrong results too? Because PLANNER tends to get it less wrong and
there's work-in-progress [1] that ORCA doesn't want to have to invent
separately.

In the meantime the workaround is to coerce a separate var. That can be
done using a temp table or view. Above query can be rewritten as:
  ```
  CREATE TEMP TABLE tempt AS SELECT i AS ai1, i AS ai2 FROM generate_series(1, 3)i;
  SELECT ai1, ai2 FROM tempt GROUP BY ai2, ROLLUP(ai1) ORDER BY ai1, ai2;
  ```

[1] https://www.postgresql.org/message-id/flat/830269.1656693747@sss.pgh.pa.us