Acquire sample rows for root partition to build extend statistics
We used to not acquire sample rows for root partition when the following conditions were held for
all attributes.
```c
	if (get_rel_relkind(attr->attrelid) == RELKIND_PARTITIONED_TABLE &&
		!get_rel_relispartition(attr->attrelid) &&
		leaf_parts_analyzed(stats->attr->attrelid, InvalidOid, va_cols, stats->elevel) &&
		((!OidIsValid(eqopr)) || op_hashjoinable(eqopr, stats->attrtypid)))
```
It works fine when we just have per-column-statistics, we can merge the statistics of leaf
partitions into statistics of the root partition. But it can't work with extended statistics. We
must acquire sample rows for root partitions to build extended statistics.

Fixed issue: https://github.com/greenplum-db/gpdb/issues/13872.