(main only): fix wrong results caused by over-eager constraint exclusion (#14553)
When checking relation exclusion by constraints, GPDB performs further
checks that if the predicate is of format expr=constant, try to evaluate
the clause by replacing every occurrence of expr with the constant. But,
this does not work for all cases.

the related issues are as follows:

remove simple_equality_predicate_refuted to fix wrong results caused by
over-eager constraint exclusion.