Fix flaky test 'partition'
ALTER DEFAULT PRIVILEGES IN SCHEMA ... affects all future tables
created in the same schema. Since the test 'partition' is running
in parallel with other tests like 'partition1', when some table is
created in 'partition1' test after the SELECT privilege on 'public'
is granted to a role in 'partition' test, that role cannot be dropped
later because that table still exists in another test, e.g.:

	 DROP ROLE user_prt_acl;
	+DETAIL:  privileges for table start_exclusive_smallint
	+ERROR:  role "user_prt_acl" cannot be dropped because some objects depend on it
	+privileges for table start_exclusive_smallint_1_prt_p1
	+privileges for table start_exclusive_smallint_1_prt_pmax

Fixing it by using a dedicated schema for the ALTER DEFAULT PRIVILEGES
test in 'partition'. Also leaving the table, schema and role not dropped
in order to test upgrade.