No need pstrdup to initialize the dispatched query string
Minor correction that we don't need to call pstrdup() when initializing
the query string. exec_mpp_query() would fill in a dummy string if it
is empty.

Noticed this when testing some specific workloads and seeing overheads
in memory allocator caused by this. Didn't dig more into the impact but
let's just fix it anyway.