Remove merge fixme from isolation2 setup test
The fixme is for pg_basebackup --no-verify-checksums flag. This
feature doesn't work and hence is not supported currently in GPDB as
no way to identify based on physical files if AO or HEAP. Need to come
up with some mechanism to enable this feature. Till that happens, no
point having FIXME related to it in test file.

Adding breadcrumbs with GPDB_12_MERGE_FEATURE_NOT_SUPPORTED comment to
fix the isolation2 setup file as well once feature is
enabled. Removing the fixme from setup.sql and setup.out file.

Reviewed-by: Huansong Fu <fuhuansong@gmail.com>