Fix bogus ORCA plan joining CTE and replicated table (#14896)
With some conditions ORCA generates Redistribute Motion from all segments to
one segment or generates a plan which can't be executed by reason of
inconsistency between CTE producers and consumers.

In such plans Shared Scan is CTE producer that's executed only on one segment
(due to Redistribute Motion from all segments to one segments), but CTE 
consumers are executed on all segments and then query hangs.

We get Redistribute from all segments to one segments because the following.
When translating expression to DXL, ORCA sets one segment for input array if
strict or tainted replicated distribution detected. This one segment is used as
output array for below motions including redistribute and broadcast.

Solution is to prohibit replicated distribution when required distribution in
sequence is singleton on master or non-singleton with prohibited replicated.

In case of inlining one CTE we also need to push down non-singleton with
prohibited replicated through filter. Because filter is appeared when CTE may 
be inlined and inlining allowed or only one consumer exists for that CTE. In
this case ORCA append logical select, which later is translated to filter