Optimize ConstructRootColMappingPerPart for common case
This method is very expensive when we have many partitions and many
columns. In a partition table, it will iterate over each partition, and
for each partition check every column against each root's column. This
is because the child partition can have dropped columns and columns in a
different order than the root. Because this is a less-common scenario,
we should optimize for the common case here. While iterating through the
root's columns, as long as the child's column also matches, we don't
need to recheck child columns that have previously been found. If there
are no dropped/swapped columns, this makes the check go from O(n^2) for
each partition to O(n).

As an example, for a 1200 partition, 200 column table of type `int` with no
dropped/swapped columns, a simple `select * from part_table` query went from 6
to 5.3s with an empty relcache, and 560s to 110ms with the table in the
relcache. With more partitions and more columns the speedup gets even
larger.