Fix IC bugs in ic_proxy_ibuf_push() (#15139)
My previous comment 57a3140 added magic number field in the ICProxyPkt. Some bugs are found in the process of porting it to 6x: they are related to the invalid value of ICProxyPkt field, current code doesn't handle them since they don't exist in the normal package.

All of them have been fixed in the 6x PR: #14998, so this PR backports the fix code and test cases into the main branch.