Assert AO/CO exclusion for anti-wraparound vacuums
Anti-wraparound autovacuums are meant primarily to update transaction
metadata and currently they are executed on user tables as well.
Executing them on AO tables is pointless from this perspective as they
don't carry any transaction metadata, and will lead to unintended cycles
spent on vacuuming their contents. Their aux tables carry it and will be
anti-wraparound autovacuumed anyway.

This is already enforced in the code, add an assert for the same.

Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>