Check partition key opfamily in partition pruning (#15260)
Check partition key opfamily in partition pruning

Issue
As of now, ORCA's constraint framework takes limited consideration for
opfamilies while deriving properties from predicates. This has caused issues
like missing motion, incorrect direct dispatch, incorrect static and dynamic
partition pruning.

Root cause
In order for the query planning to take advantage of the existing data
distribution or partition, i.e. avoid unnecessary motions or partition scanning,
we have to ensure the predicate's operator belongs to distribution or partition
key's opfamily. When this check is missing, we could end up with incorrect plans
and results, by failing to trigger necessary motions, or scan required
partitions.

Solution
The core idea is, compare the operator to the distribution key's opfamily when
evaluating a hash join condition; and compare the operator to the partition
key's opfamily when evaluating a partition filter.

Implementation
The implementation is much more pervasive than the summary above. This is
because the same predicate is evaluated many times for different purposes in
query planning. This PR addresses four areas:

Building equivalence class to optimize join order and join conditions
(refactoring PR Fix incorrect result from hash join on char column #14986)
Explicit cast of hash join condition
Extract predicates on partition keys, used in dynamic partition pruning
Deriving interval from const predicate, used in direct dispatch and static
partition pruning
As of now, the metadata only has information on a column's distribution
opfamily. So our first step is to include the partition opfamily. For ease of
review, the PR is split into three commits. It will be merged as a single
commit.

Derive column's partition (btree) opfamily
Compare the operator to the column's opfamily. Depending on whether the
predicate is used as a hash join condition, or a partition filter, we look at
the column's hash or btree opfamily.
Tests
Future work
This PR addresses a few queries that failed for not considering the distribution
or partition key's opfamily in logical planning. We expect to find similar
errors in the constraint framework and scalar expression evaluation, and we will
address them case by case. Going from logical to physical planning, we need to
keep checking opfamilies to not miss critical motions.

Fix #14887
Fix #14982

---------

Co-authored-by: Jingyu Wang <wjingyu@vmware.com>