Exclude views from missing file view check
We could create a view out of a table via the ON SELECT rule on the
table. That leaves a view with valid relfilenode, but its relation file
might've been cleaned up. Now do not show it when checking missing files.

P.S. latest upstream has removed this feature (https://github.com/postgres/postgres/commit/b23cd185fd5).
We considered backporting the same but eventually decided that it is
too early to remove this feature which might have bigger impact to GPDB.