Fix coordinator crash in MPPnoticeReceiver (#15664)
 fix issues #15194 

We got an unexpected msg while idle. So, we construct a simple notice msg use pqInternalNotice
which do not specify the line no. Due to the line is NULL we got segfault in MPPnoticeReceiver while
copying it to QENotice using strlcpy(notice->line, line, sizeof(notice->line));

Solution: use the same way as file does. Use SIZE_VARLEN_FIELD and COPY_VARLEN_FIELD instead of strlcpy.
SIZE_VARLEN_FIELD and COPY_VARLEN_FIELD will process NULL value.

In MPPnoticeReceiver the initialize value of file and func is "".
```
char	   *file = "";
char	   *func = "";
```
If the notice we received does not contain file or func info, file or func will still be "".
In SIZE_VARLEN_FIELD and COPY_VARLEN_FIELD we only judge NULL value not
judge "" value, so the length will be calculated as 1 and "" will be copied to QENotice,
but this has no meanings.

In forwardQENotices if file is NULL we do not send file info to client.
however, here we send "" to client.
```
if (notice->file)
{
	pq_sendbyte(&msgbuf, PG_DIAG_SOURCE_FILE);
	pq_sendstring(&msgbuf, notice->file);
}
```

So change initialize value of file and func in MPPnoticeReceiver to NULL to avoid sending unnecessary
"" to client.

backport from 6x: https://github.com/greenplum-db/gpdb/pull/15397