Resolve two FIXMEs in extractPageInfo()
There are three points in the two FIXMEs. Let's review them one by one:

1. Should we return for RM_APPEND_ONLY_ID: confirmed that RM_APPEND_ONLY_ID
doesn't have any registered buffers in all cases. But I don't think we need
to return there because the `for` loop will just be a no-op. There are
other types of WALs that don't have registered buffers either, and we do
not need to treat AO ones specially.

2. Should we take care of RM_DISTRIBUTEDLOG_ID: I don't think we need to
because it should be treated in the same way as for other SLRU related
WALs. Also, there's no buffer registered for it either.

3. Make sure RM_BITMAP_ID works after we took the upstream WAL format revamp
(commit 2c03216d8311): I think the current code looks fine and we do not need
to do anything. Basically, we used to handle bitmap and some other WAL types
in extractPageInfo specially because we needed to decode the data portion in
the WAL first and get the block number from different corresponding structs.
With the revamp we store block number in standardized format and decoding it
becomes the same way between different WAL types. Bitmap WALs do not look to
be any different.
Also one more issue that's not mentioned by the FIXME is that, for bitmap we
used to handle a different forknum than the MAIN_FORKNUM which is INIT_FORKNUM
in some cases (stored in xlrec->bm_fork). Now we always only handle MAIN_FORKNUM
in extractPageInfo() like any other cases. However, this is fine because all
forks other than the MAIN_FORKNUM is going to be ignored in process_target_wal_block_change()
anyway because there won't be the same source data file to be found (see comments
for isRelDataFile()). I think this was also what happened when we tried to call
pageinfo_add() for INIT_FORKNUM before the revamp. Now we just explicitly do
not do that for non-main forknums.