Support executing foreign tables from segments in Orca (#15559)
When the mpp_execute option "ANY" is specified, we now will either
execute it on a segment or coordinator depending on the cost.

Previously, when a foreign table was created to be executed with the ANY
location, Orca still only executed it on the coordinator. Now, we treat
it similarly to executing a set-returning function such as
`generate_series()`, and allow it to be executed on either the coordinator
or a segment depending on cost. We add a new table distribution,
Universal, which derives a Universal distribution in Orca just like a
generate_series. It's important to note that this has different behavior
from `all segments`, which treats the data as distributed randomly and
processed as separate data on each segment (for example, accessing a
specific file on each segment).

As an example, if we're using the postgres FDW, for a simple
`select * from pg_foreign_table`, we should perform this on the coordinator
since it would end up being gathered to the coordinator after. However, if
we're performing a join, it's (very) likely to be better to do on a
segment.

Additionally, this PR adds the gp_postgres_fdw regression suite to ICW.