Fall back to planner for queries with foreign partitions using greenplum_fdw (#15706)
This fdw has special logic that populates fdw_private and performs
parallel cursor setup on the coordinator in `BeginForeignScan`, which is
called in `ExecInitForeignScanForPartition`. Within the Dynamic nodes,
Orca calls this function during execution for each partition, as there
is partition-specific logic that needs to be done to move to the next
partition. Planner on the other hand does this during initialization for
each individual partition, not during execution.

There may be a way to refactor this logic to be performed on the
coordinator or expand dynamic scans into individual scans similar to
planner, but for now let's fall back to planner to prevent a crash.
Note that Orca still supports the greenplum_fdw if it is not a
partition.