Add missing keywords to PartitionIdentKeyword
As the FIXME suggests, many keywords added during the PG12 merge are not
added to PartitionIdentKeyword. Add them now, so that we could use them
as column alias without a preceding `AS`, or as partition name.

What's not covered: keywords that are in both 6X and 7X. I didn't go over
all of the keywords that are not in PartitionIdentKeyword.

P.S. it is actually a GPDB-specific feature that we allow keywords to be
used for column alias without `AS` (commit a64af2d78ebc). In upstream only
starting from PG14 this has been supported (commit 06a7c3154f5). We would
be adapting to that method when we catch up with PG14.