[ORCA] Update index cost model to account for INCLUDE columns (#15697)
After commit b9f43b93, ORCA supports cover INCLUDE index columns. But
the cost model does not yet account for those columns. An INCLUDE column
is extra payload embedded into the physical index file. That requires
extra I/O to read tuples from the index that is proportional to the
width of the INCLUDE columns. In short, ORCA should always perform
index-only-scan using the index with the narrowest width that satisfies
the query.

For example:

  ```
  CREATE TABLE t1 (c1 int, c2 int, c3 int, c4 int, c5 int);

  CREATE INDEX idx_large ON t1 USING btree(c1) INCLUDE (c2, c3, c4, c5);
  CREATE INDEX idx_c1 ON t1 USING btree(c1);

  EXPLAIN ANALYZE SELECT c1 FROM t1 WHERE c1= 4;
  ```

Given that both indexes completely cover the predicate and project
columns, it is always better to use the index with the smallest payload.
In this case, that's idx_c1 (idx_large contains columns c2, c3, c4, c5
that are unused in the query whereas idx_c1 contains 0 unused columns).

An additional related change is to update the usage of relallvisible
fraction to cost index-only-scan more correctly. Previously, the value
was multiplied by a constant representing the random I/O cost to
retrieve tuples from the heap that are referenced from the index. That
sort of make sense, because if heap is all visible then no heap accesses
are required which presumably meant no random page accesses. However,
depending on the index implementation, I think there may still be random
page cost in reading the index itself.

For now, remove the previous usage because some ANALYZE'd regress tests
suggest it still needs to be accounted for, although maybe via a custom
constant. Instead, use relallvisible to proportionally calculate the I/O
cost of accessing the whole row from a heap as opposed to just the index
itself (which may be narrower). That should provide a more accurate I/O
cost.

In order to more clearly define the cost formula a new cost parameter
EcpIndexOnlyScanTupCostUnit is added to distinguish between the scan
cost of reading both the index and heap (EcpIndexScanTupCostUnit) as
opposed to just the index. For now they are the same values, but we can
tweak in the future as needed.

Really though, these cost unit parameters are a bit deceiving. The
reason that wider indexes/tables are more expensive is not so much due
to the IO of reading a wider row from the page. Instead, it is likely
due to the subsequent effect of needing to page more blocks from disk in
and out of memory. Unfortunately, these parameters do not really take
that into account.

Intent of this parameter is to be a stop-gap until we create a new cost
model that directly takes pages into account, similar to planner (e.g.
cost_index).