function_extensions: Fix tmpdir ls flake
The function_extensions test used to flake as follows:

-- Temp file number before running INITPLAN function
 SELECT get_temp_file_num();
      get_temp_file_num
 ------------
- 0 files and 0 directories
+ 0 files and 1 directories
 (1 row)

It is possible that some temp directories are present sometimes before
the we execute the INITPLAN function, just like this old comment
suggests that temp dirs like pgsql_tmp3889679.0.sharedfileset may hang
around:

-- Temp file number after running INITPLAN function. All the files should've
-- been cleaned up, but it's normal that the temp directory to hold them is
-- still around.

So, in this commit, we don't even bother about the directories, we just
look at the files. That too, we also eliminate any other possibility of
flakes by doing a before-and-after-equality-check.

PS: Tried to use pg_ls_tmpdir(), however it doesn't recurse into
SharedFileSets, as this thread suggests:
https://www.postgresql.org/message-id/20191227170220.GE12890%40telsasoft.com