Consider bitmap alternative only for ANY ScalarArray
Commit 290166f introduced ScalarArrayOpExpr as an indexable qual for ANY
operator only as it is translated into an OR combination of indexscan on
the array elements.  Prior to this commit ORCA considered Bitmap
alternatives for equality on ScalarArrayCmp (both ANY and ALL) whereas
similar to planner, ORCA should also only consider it for equality on
ANY op. This commit fixes that issue and adds tests for the same.