Open parallel cursors on behalf of the current user (#15820)
It's a weird behavior that parallel retrieve cursors are opened on
behalf of the session user no matter who the current user is, and it
confuses the database users.

This commit changes the behavior to open parallel retrieve cursors on
behalf of the current user, just as how these parallel retrieve cursor
functions check permissions.