Refactoring RTE deduplication logic from PR #14304
Add comments for result relation RTE deduplication logic.

One major change is we only update the RTE look up table after appending a new
RTE to the range table entries list. We used to update the look up table when
user requests the RTE index within the entries list. The problem with that is
theoretically user can just request the RTE index without actually updating the
entries list. In that case, if we prematurely update the look up table, the
assigned query id could eventually point to a non-existent RTE or wrong RTE.