fix bitmap index (#15923)
This PR is trying to fix the issue of #13517. The RCA could be found at huansong's comments.

The data registered by XLogRegisterBufData() could be replaced by Full Image, and the value
of bm_next_blkno in special data is page data too, if we change that, we should record it in XLog,
instead of predicting it in the replay process.