Remove a FIXME in cdbaocsam.h
The FIXME reads quite confusing at first: it is placed between two things that
are totally unrelated to bitmap. It is not clear what exactly is the "new method"
that is supposed to have another MERGE_FIXME, and what exactly should we store
in the "new struct".

When looking at the original commit that introduced it: fb7f49c1a20b61f079614dd44aad1153ff723e72,
it becomes somewhat clearer that the original purpose of this FIXME might have
been actually addressed by the same commit. The "new method" it talks about is
the scan_begin_extractcolumns_bm() handler, which had a MERGE_FIXME at that time:
https://github.com/greenplum-db/gpdb/commit/fb7f49c1a20b61f079614dd44aad1153ff723e72#diff-bfe700fba9b92e12fb7e0d77101d4bef63591973e55662f6fffdf66e2333d919R216-R219.
And the "new struct" is the AOCSBitmapScanData which the commit took out from
AOCSScanDescData for bitmap-related things. Therefore, this FIXME message seems
to be left there by mistake. Remove it now.