Invalidate its foreign table's relcache when alter server
Foreign table's distribution policy depends on the foreign server's
`mpp_excute` and `num_segments` options, to force update the cached
policy, if the option matches, invalidate the relcache of this server's
every foreign table.

If not, Greenplum still provides the old policy which is stored in the
relation cache since the table's catalog never changed, the generated
plan would be wrong.