Do not use immediate restart in regress test
There is nothing wrong with doing that for the tests themselves, but it has an
implication to a bug #16120 where restarting w/o checkpoint can create orphaned
files and results in the failure of test gp_check_fiels. While that issue is
dealt with separately, for now turn the restart mode to fast to avoid the issue.

We ignored isolation2 tests for now because gp_check_files are not covering them
(and it would be really hard to cover, given so many intentional crash/panic or
other ad-hoc scenarios that we are doing in isolation2 could create orphaned
files... I don't see us do that anytime soon).