[ORCA] Enable index-only scan on AO tables (#16162)
In the past, ORCA and PLANNER had disabled index scan on append only
(AO) tables due to poor performance. The reason is due to the following
design decisions of AO tables:

  * AO tables have a variable block size layout on disk (e.g. batch
    insert creates larger block than single row insert). However,
    this makes index scans tricky because the block identifier does
    not directly map to a fixed offset in the relfile. Instead, an
    additional abstraction layer, the block directory, is required
    to map the block identifier to an offset inside the relfile.

  * AO table blocks are loaded in-memory into a single varblock.
    Heap blocks, however, support multiple in-memory instances and
    fit in the page cache. That means random I/O on AO tables is
    more susceptible to thrashing.

  * AO tables in production often compress the blocks. That adds an
    additional penalty for loading blocks that is exacerbated by a
    poor page replacement algorithm. And, in the case of AO single
    varblock, any time we have to revisit a block, it will always
    have to be replaced and reloaded.

Index-only scan is able to mitigate some these performance barriers by
skipping block access. In those cases, we allow ORCA to generate an
index-only scan alternative on AO tables.

NOTE: Executor does not support index-only scans on an AO table that was
upgraded from a pre-GPDB 7 version. In that case, ORCA will not generate
an index-only scan alternative.