Fix redundant sort being enforced on group aggregate (#16276)
**Issue:**
Redundant sort is being enforced on group aggregate.

For a Query like,
`EXPLAIN (costs off) SELECT a FROM foo GROUP BY a ORDER BY a;`
Plan generated before fix:
                QUERY PLAN
------------------------------------------
 Gather Motion 1:1  (slice1; segments: 1)
   Merge Key: a
   ->  Sort
         Sort Key: a
         ->  GroupAggregate
               Group Key: a
               ->  Sort
                     Sort Key: a
                     ->  Seq Scan on foo

Sort operator enforced on top of GroupAggregate is redundant
as output from GroupAggregate is already in sorted order.

**RCA:**
If num_of_rows to be sorted is estimated as <= 1, cost is estimated as 0.
The Sort expression is being considered first
for finding an expr that satisfies the optimization context with sort requirement.
Since it has 0 cost, it is redundantly added on top of aggregate operator.

**FIX:**
To avoid costing the Sort operator as 0 when num_of_rows <= 1,
minimum num_of_rows is being set as 2 just like it is done in postgres optimizer.
This is fetching a minimum cost value of > 0, and sort won't be added redundantly.

**Test related changes**
* Fix /src/test/regress tests plan diff failures
* Fix failing mdp tests because of plan changes
* Could not fix DQA-SplitScalarOnDistCol.mdp test case with fix_mdps.py,
so had to add new mdp file from command in test comment.
* Add new mdp test case for issue query.