Correcting derived distribution spec for CPhysicalJoin. (#16342)
Problem: Wrong results generated for subquery in projection list for replicated
tables.

Analysis: To derive distribution for any join operator,
CPhysicalJoin::PdsDerive() is invoked. For deriving distribution it checks
DistributionSpec for outer and inner children. when we have DistributionSpec for
outer child as replicate and inner child  as universal then we return universal
as derived distribution. Eventually "Gather Motion" is not created and as data
is not there with coordinator so it gives no rows as output.