Fix "cache lookup failed for foreign table" for multi-level partition table with foreign tables in Orca (#16376)
This issue was introduced by #15706. When ORCA translates multi-level
partition table Relcache to DXL, it searches for the storage type of
the leaf table and checks if the foreign table is a greenplum_fdw table.
However, the none-leaf table is also checked, and an ERROR is thrown
because the none-leaf table is not a foreign table.

Fix this issue by checking if the table is a foreign table first to avoid
elog ERROR.

Co-authored-by: wuyuhao28 <wuyuhao28@github.com>