Compute statistics for a column of table (#16487)
Added logic to compute statistics for a column of table

MOTIVATION
----------------------------
While creating 'Physical Scan Opertors', the statistics of base table are computed
in function CPhysicalScan::ComputeTableStats().But while computing these statistics,
the histograms for the columns of table are not generated as the columns of the
table are not sent to CMDAccessor::Pstats(,,,,).

While costing (Dynamic)Index(only) scans, NDV of columns of table is required.
The NDV cannot be calculated as the histograms were not generated during creation
of scan operator.

SOLUTION
----------------------------
To overcome this, function 'CStatistics::ComputeColStats (,,) is added.
This function computes the statistics for a given column. The generated statistics
are used for (Dynaic)Index(only) scan costing.

NOTE
----------------------------
This 'fix' updates the logic required for 'computation of statistics for unused index column'
required in the PR #16156 on Index costing.

Co-authored-by: nishant sharma <nishants4@nishants46M6RY.vmware.com>