Fix the type oid that left type and right type are not the same
min-max filter assumes that the left type is the same as the right type.
It's not truth in some cases. For example:
where d_date between '1999-4-01' and
  (cast('1999-4-01' as date) + '60 days'::interval)

The left type is date, while the right type is timestamp in the second
comparison. The comparing functions are not the same.

NOTE: The right type is kept in ScanKeyData.sk_subtype.