Add support for min/max aggregates optimization (#16480)
* Add support for min/max aggregates optimization

This commit adds support for min(), max() aggregates optimization, going
forward ORCA uses IndexScans(Forward/Backward) along with a limit operator to
perform min(), max() on a btree index key. Prior to this commit, ORCA uses
regular aggregates with Sequential Scan to accomplish this. Along with this
optimization, this commit also adds support for IS NULL and IS NOT NULL
conditions on a btree index column and uses an Index Scan for such predicates
instead of a Sequential Scan. This support was required(at least for IS NOT
NULL condition) in order to proceed with min/max optimization. New plan changes
are quite straightforward with an IndexScan and a Limit operator instead of
aggregate with sequential scan. Leveraging index scans with limit could improve
the query performance significantly.