[ORCA] Add CLogical[Dynamic]IndexOnlyGet operators (#16671)
Prior to this commit, CLogicalIndexGet represented both index and index
only operators. It was not until implementation phase that a physical
index only scan operator was considered. The issue with delaying that
decision is that when hints are introduced, it would force the logic to
reside in the implementation phase. Instead, we want to enable or
disable the hinted alternatives earlier, during the exploration phase.

By separating the index and index-only operators in the exploration
phase, the hint logic for index and index-only scans can also reside in
exploration phase.

An epsilon cost is added to break index and index-only ties. After
adding a separate logical index-only-scan operator, some regress test
plans started unexpectedly generating index-only-scans plans (e.g.
qp_indexscan and aggregates). Some debugging revealed that index-scan
and index-only-scan plans in these cases were costed the same. This
meant that the order that the alternatives were generated effected the
plan.

Instead, ORCA should distinguish between the two plans. The reason that
the plans were costed the same is that 0% of the tables' blocks were all
visible (i.e. heap access was required). In that case index-only-scan
has no advantage and in fact can be slightly worse because it still
requires a bitmap visibility lookup. This commit also adds a visibility
map lookup cost to prefer index-scan in this case.