Fix LeftJoinPruning pruns essential left join (#16690)
ORCA performs the left join pruning at the preprocessing stage. For performing
the pruning operation ORCA creates a set of columns which are output by the
child operators. In the current implementation to generate the output columns
for the child operators following steps are performed
1.) Derive output columns of the logical parent operator
2.) If scalar operator is present
	2a.) Derive used columns of the scalar operator
	2b.) Create a set combining columns from step 1 and 2a
But for some operators like UNION where no scalar child is present ORCA was
creating a empty set of output columns for the child operators which was
leading to wrong pruning of left joins. This PR fixes that issue by modifying
the logic for computing output columns.
Authored-by: tongshu <wangpan.wang@alibaba-inc.com>