Derive dynamic table scan cardinality from leaf parts
**Motivation**

Currently, ORCA lacks support for deriving stats for predicates that involve
dates and other time-related data types. This limitation arises from the
internal representation of these data types -- date as long int whereas other
time-related types as doubles. This design choice was made for implementation
convenience. Integers provide sufficient granularity for date types, and doubles
are necessary for other time-related types to accommodate precision down to
microseconds. However, this choice prevents the stats derivation for predicates
involving different data types. Direct comparison is impossible when the column
is represented by int and the value by double. As of now, ORCA detects such
predicates and uses default parameters to calculate stats. This often results in
inaccurate stats and poor plan performance. This PR alters the internal
representation of date types to doubles as well, and enables stats derivation in
predicates involving different time-related data types.

**Implementation**

Converting the date type to double is a straightforward transition. ORCA already
supports stats derivation for data types represented by doubles. The complexity
arises from ORCA's use of the constraint framework, where predicates are
converted to constraints. Previously, when the date was represented by int,
individual dates were isolated points with no reference to specific time. An
equality predicate of the date type translated into a singleton constraint. Now,
with the date type represented by double, the same equality predicate is no
longer a singleton but a range constraint. It compasses a lower and upper bound
presenting the start and end of the day.

This internal data type transition may introduce discrepancies in stats
derivation through bucket sampling. Similar to constraints, histogram buckets
have lower and upper bounds. Whereas a singleton constraint can fall into one
bucket only, a range constraint will likely span across multiple. This kind of
inconsistency may results in inaccurate stats and should be mitigated. One
notable example is the computation of tuples to be scanned in static partition
elimination. This is currently achieved by comparing the predicate constraints
on the partition key to the histogram buckets. The comparison dervies a scale
factor, which is then used to calculate the filtered tuple count. To address the
aforementioned issue, this PR adopts a more robust approach by directly summing
the tuple count from the leaf partitions slated for scanning. This is
facilitated in this version of Greenplum through the storage of leaf partition
stats in the metadata.

**Discussions**

In addition to routine CI tests, this PR has undergone testing in both explain
and workload pipelines. Given that the code update focuses on stats tuning,
there is less concern about it leading to incorrect query results. The primary
focus was on identifying factors that could impact plan performance.

We observed variations in plans involving hash vs. group-by aggregations. This
is expected, considering this PR's purpose is to derive more accurate
cardinality, which often influences the optimal choice of aggregation
operations. We also noticed changes in join orders involving multiple partition
tables. This can be attributed to how ORCA estimates cardinality in dynamic
partition elimination. In a join where both tables can act as the partition
selector for the other, the cost model may favor a specific join order if the
partition selector can sufficiently reduce the outer cardinality to offset the
higher cost of hashing a larger inner table. It's essentially a trade-off
between which side is more suitable for hashing versus being the selector side
of DPE.

It is crucial to verify that the cardinality of (dynamic) table scans involving
single-type predicates remains mostly consistent compared to the previous state.
However, in the upcoming tasks, we should limit the above mentioned discount
effect by DPE if static pruning already eliminates a majority of the partitions.