Align broadcast penalty in Orca's DPv2 to guc value
In the join enumeration algorithms, we use a simplified cost model.
However, we still want to align it as much as possible to the cost model
used in optimization. In that cost model, we penalize broadcasts if they
exceed optimizer_penalize_broadcast_threshold, and thus we need to do
the same here.