pg_upgrade: rewrite data type check from sql to plpgsql to run on GPDB6
pg_upgrade has a few checks that look for specific data types. The SQL
used to look for data types does not run on GPDB6 because it is a
recursive CTE that contains a self-reference in a subquery. This
specific type of query works on GPDB7, but is broken on GPDB6 and
therefore doesn't work for GPDB6 > GPDB7 upgrade. To get type of query
working on GPDB6 would require a non trivial amount of backports. As a
workaround, the query was rewritten in plpgsql so it can run on GPDB6.

This fixes the following pg_upgrade checks:
 - check_for_composite_data_type_usage
 - check_for_reg_data_type_usage
 - old_11_check_for_sql_identifier_data_type_usage
 - old_9_6_check_for_unknown_data_type_usage