Fix race condition in CTE reader-writer communication (#16431)
Previously, a race condition between Share Input Scan (SISC) reader and
writer could cause the writer to hang if the reader node was squelched
before the writer node was initialized.

Here's an example plan:

EXPLAIN (COSTS OFF)
WITH cte AS MATERIALIZED (SELECT * FROM sisc_t1)
SELECT y.a, z.a
FROM (SELECT cte1.a, cte1.b FROM cte cte1 JOIN sisc_t2 ON (cte1.a = sisc_t2.d)) y,
     (SELECT cte2.a, cte2.b FROM cte cte2 JOIN sisc_t2 ON (cte2.a = sisc_t2.d)) z
WHERE y.b = z.b - 1;

                                        QUERY PLAN
------------------------------------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)
   ->  Sequence
         ->  Shared Scan (share slice:id 1:0)
               ->  Seq Scan on t1
         ->  Hash Join
               Hash Cond: (share0_ref3.b = (share0_ref2.b - 53))
               ->  Redistribute Motion 3:3  (slice2; segments: 3)
                     Hash Key: share0_ref3.b
                     ->  Hash Join
                           Hash Cond: (share0_ref3.a = t2.d)
                           ->  Shared Scan (share slice:id 2:0)
                           ->  Hash
                                 ->  Redistribute Motion 3:3  (slice3; segments: 3)
                                       Hash Key: t2.d
                                       ->  Seq Scan on t2
               ->  Hash
                     ->  Redistribute Motion 3:3  (slice4; segments: 3)
                           Hash Key: (share0_ref2.b - 53)
                           ->  Hash Join
                                 Hash Cond: (share0_ref2.a = t2_1.d)
                                 ->  Shared Scan (share slice:id 4:0)
                                 ->  Hash
                                       ->  Redistribute Motion 3:3  (slice5; segments: 3)
                                             Hash Key: t2_1.d
                                             ->  Seq Scan on t2 t2_1
 Optimizer: GPORCA
(26 rows)

In this plan, both Share Input Scan reader nodes on slice2 and slice4
on seg2 are squelched due to no result from the inner sides of the
Hash Joins. The squelching could happen even before the Share Input
Scan writer node on slice1 is initialized.

For cross-slice shared scan, readers and writer communicate through
shareinput_Xslice_state (xslice state in short) in shared memory.
There is a hash table shareinput_Xslice_hash in shared memory that
contains one "xslice state" entry per shared scan. Currently, the
"xslice state" entry is created by the first participant that
initializes, which can be a producer or a consumer; It is removed by
the last participant that references it, which also can be a producer
or a consumer.

In most cases when all participants are executed, the producer would
be the last one who references the "xslice state".  However, in case
of reader initialized and squelched before the writer initializes, the
reader would remove the "xslice state" entry from the hash table even
before the writer references it. So by the time the writer initializes,
it will be ignorant of the already removed "xslice state" and
create a new one. In this case, the write wish to communicate with
the readers through the new "xslice state", while the readers are
already squelched. So the writer will hang.

This patch fixes the problem by disallowing the reader to remove the
"xslice state" entry from the hash table if it is squelching.

Add GUC debug_shareinput_xslice for finer grained logging.

Co-authored-by: wenru yan <ywenru@vmware.com>
Co-authored-by: Lei (Alexandra) Wang <alexandra.wanglei@gmail.com>

fix bfv_cte