Create pg_plan_hint parser hook
Both ORCA and PLANNER support the same query hint syntax. So naturally,
it would be best if both could leverage the same parser. For PLANNER the
parser is implemented inside the pg_hint_plan extension. This commit
defines a hook that calls the parser and turns the Query into HintState
structures.