Include distributed xid in transaction commit WAL in all cases
Currently, we include the distributed xid in the transaction commit WAL only in
these two cases:
1. For 1PC, QE writes it at the end of the commit;
2. For 2PC, QD writes it after PREPARE phase is done on all segments.

However, we are missing a case where for 2PC, QE writes a transaction commit WAL
at the end of COMMIT phase. We are not including distributed xid in that record.
Therefore, a segment replaying that record won't initialize distributed log for
that commit.

This has not been an issue because when a QE needs to replay a 2PC (that is,
only during recovery), it does not need to conduct dtx visibility check for the
2PC, since no query is allowed during recovery.

However, this has become issue when we want to support hot standby or change
data capture (CDC) where query is allowed during continous recovery.
Therefore, we now include the distributed xid for any transaction commit that
has a valid distributed xid, which fills the missing piece as mentioned above.