pg_upgrade: check for views using removed operators
Views that use removed operators will cause upgrade to fail. This
happens during metadata restore on the target cluster because pg_restore
will error try to create a view using operators that do not exist
anymore. This is not ideal as we could be many hours into upgrade before
a view using a removed operator causes pg_upgrade to fail. This check
calls a support function on the source cluster to check if views using
removed operators exist.