pg_upgrade: modify `--output-dir` to not create certain sub-directories
Previously when specifying `--output-dir`, pg_upgrade creates
`<output-dir-option>/pg_upgrade_output.d/<timestamp>/[files and logs]`.

This made it very hard to manage the directories when wrapped with
gpupgrade. The timestamps are created on the fly, which made each
pg_upgrade have unique folder. Unfortunately, this made it hard to track
down the pg_upgrades that ran with gpupgrade. Ideally, all pg_upgrade
that are run on the segments have the same timestamps.

Modify `--output-dir` such that we only need to give the directory we
want to put the files and logs. The user of pg_ugprade will be the one
responsible for creating unique log directories.

With this change, pg_upgrade will create
`<output-dir-option>/[files and logs]`.

The data directory already has a log directory which conflicts with
`pg_upgrade`. So by default (without the `--output-dir` option flag),
`pg_upgrade` will still create
`<data-directory>/pg_upgrade_output.d/<timestamp>/[files and logs]`.