Fix issue https://github.com/greenplum-db/gpdb/issues/17333.
In the adjust_setop_arguments(), if the setop is partitioned and subpath locus is general,
it will call make_motion_hash_all_targets(), and this function will change the locus to singleQE
when none of the columns are hashable. But mark_append_locus() will change it back to the original
partitioned locus, which will result a wrong Motion node added in the top.

This PR moves check for non-hashable column(no targetlist) to choose_setop_type().