Inherit parent's options for child like gp-style partition table (#695)
Currently, the child partition doesn't inherit the table access
method, column encoding and reloptions from its parent table
when the child table is added later. The change causes
in-compatible behavior for users from gpdb.

To achieve the target, the kernel saves table access method in
pg_class for gp-style partition table, to be inherited by the
child table.
Reloption and column encoding of the child partitions also inherits
from its parent if they are absent.

However, the bahaviors between greenplum and cloudberry are still
not 100% compatible.