Avoid REFREH fast path if matview has foreign tables.
For matview has foreign tables, we never know the data
status.
Since we usually create a matview has read external tables,
the initial data status is Up-to-date.
That will make REFRESH fail to do the real thing with
fast path feature, the worst case is we never try to get
the external data during REFRESH.

Authored-by: Zhang Mingli avamingli@gmail.com