Fix: PAX fast sequence will return the same seq in concurrent case
The PAX fast sequence may get the same sequence in concurrent case. Because
the lock mode in the pg_ext_aux.pg_pax_fastsequence is RowExclusiveLock which
 is not an exclusive lock for query.

But if we use an exclusive lock(for query) on table pg_ext_aux.pg_pax_fastsequence,
 performance will become a problem. So in current change, added a CTID lock
to avoid getting the same sequence.