Adapt dynamic (bitmap/index/table) scan to PG 14
In PG commit 36d22dd, if the ops in tableslot which returned by
the current execution operator is fixed, the executor will simplify
the call and directly call `*virt` functions. But in the dynamic
(bitmap/index/table) scan operator, we don't know the tupleslot
ops of the subtable (it may even be different).

So the current change sets the `opsisfixed` of scan and result to
false, which can prevent the executor from directly calling `*virt`
funcstion.