Fix "unexpected gang size" issue.
In postgres planner, after top_plan having been produced, we have to
adjust the Motion type of subplans, which call cdbllize_decorate_subplans_with_motions().

If subquery has another subquery, for example, in this test case, the second subquery's
motion type is set to Gather but should be Broadcast.  The root cause is that we forget
to change local subFlow when its subplan has changed.