ORCA: Fix missing the aggstar in aggref
In the ORCA optimizer, `aggref->aggstar` is not filled.
It will cause the executor to treat `count(*)` as `count()`.

Although the results of `count()` and `count(*)` are the
same. But without `aggstar=true` may affect the custom scan.
Assuming that the storage engine pre-calculates the content
of `count(*)` and uses `aggstar` to decide whether to use
custom scan, the ORCA plan will cause the custom not work.