Fix: Core happend when calling pg_relation_size on root partitioned table with PAX AM (#1128)
In the `calculate_relation_size` method, whether to directly callback the
table AM(access method) is determined by `RelationIsNonblockRelation`.
However, for root partitioned tables, the table AM is always be NULL.

When the AM of a root partitioned table is PAX, `RelationIsNonblockRelation`
always returns TRUE. But for AO tables, `RelationIsNonblockRelation` will
returns FALSE. It because AO table are determined directly through `rel->rd_amhandler`
(PAX as a plugin, cannot register its `amhandler` in the CBDB kernel).