Fix: wrong aggno and aggtransno generated in ORCA
This issue caused `AggRef` in the `targetlist` no longer share
the same `aggno` and `aggtransno`, which means AggRef would be
calculated repeatedly.

ex.

```
	select
		sum(a) <- aggref1
	from
		any_table
	having
		sum(a) < 100; <- won't use the result of aggref1 to do the filter.
```

This problem was introduced when cloudberry cherry-picked PG14.
But after cloudberry was open-sourced, the history of the changes
had been merged, and this problem only exists in ORCA.

In PG14, it no longer computes shared AggRef in `ExecInitAgg`,
but moved the logic into the planner. However in ORCA, it only
uses `idx++` to generate aggno and aggtransno. I do think that
is a workaround logic.

The current commit fixes this behavior by adopting the same
logic as the planner. But it's worth noting that, in multi-stage
agg, the final agg often gets an outer reference `Var` in AggRef,
so for the final agg, it still cannot get the same aggno (this
logic is consistent with greenplum; in the TPCDS testing, even i
generating the same aggno for final agg, it didn't bring performance
improvement). But if necessary, we could consider moving the logic
from `dxltoplstmt` to the pre-processor, which could also reduce
some output columns.